docs: recommend check-mode for fallback pre-commit hook (#39)
As fix-mode does not return a non-zero status, it will not block submission when it detects formatting issues. Instead, it will reformat files but not stage them within `git`. The result is that users are not barred from submitting malformatted code; when they do, they'll have the fixes as unstaged changes in the files they just submitted.

Check-mode's behaviour will block submission, which is more similar to what `pre-commit.com` does and is probably what most users will expect.