fix: in use_exit_code mode, print reports to stdio
Previously we always directed the tools output into a report file, but when you ask for the exit code to be used, this means the Bazel build fails without printing the result of the tool.
Instead, all the logic related to directing output into the report file should only be used when exit code is ignored.