fix filename pattern matching
git ls-files seems to match against filename strings exactly, and
doesn't use the same logic as e.g. gitignore. this means that to match
files by exact filename below the root, we need a */ at the start of the
pattern to match more than files strictly at the root of the tree. these
prefixed patterns however miss the root, so we need to include both
forms.

additionally, the pattern array needs to be quoted to prevent the shell
from doing glob expansion against any files in the current directory.

additional edits to make the mirror_linguist_languages script work with
a mac userland.

Fixes #117 and #122.