fix filename pattern matching
git ls-files seems to match against filename strings exactly, and
doesn't use the same logic as e.g. gitignore. this means that to match
files by exact filename below the root, we need a */ at the start of the
pattern to match more than files strictly at the root of the tree. these
prefixed patterns however miss the root, so we need to include both
forms.

this is doable with bash array regexes to avoid changing the pattern
format to work with both find and git ls-files at the same time.