refactor: use rules_multitool to give defaults for formatter tools
So far this just replaces the most obvious subset of formatters that have
four binaries for Mac/Linux x86_64/arm64.

The pattern can be extended to provide more tools by default.
We should NOT try to continue the pattern to anything that requires additional Bazel rules to provide a runtime, such as Prettier.