feat: remove golangci-lint (#207)
In preparation for 1.0, we want all linters to work correctly.
In the discussion for #129 we decided that is a fatal bug, and no one is able to fix it in the short term.

We're happy to add Go linting back to rules_lint when someone has a correct implementation that handles transitive srcs.
The code will still be here in the git history when it's time to revive it.

Closes #129