Pass disable_git_attribute_checks via env instead
This fixes the ability to run `bazel run //:format -- <filenames>` with
git attribute parsing disabled.

`format.sh` expected this flag to be the last argument, however when
running with `rules_multirun` all arguments passed at the end of `bazel
run` get appended to the `args`. So, the flag that was expected to be
last was no longer last, and it re-enabled gitattribute parsing and
caused an issue with `find` not expecting this flag.

This patch makes it an environment variable, like all other information
that's passed to the format.sh script.