Swap diff tool for git
go-gitdiff which aspect CLI uses cannot support patches generated by
BSD diff which is the default on MacOS (see tagged issue).

I am investigating using ape but it requires Bazel 7.4.1 which may be
a breaking change.

Alternatively, this changes diff -> git with a few extra "hacks".
Namely, we need to replace the labels to be the same file and resolve
the symlinks otherwise `git diff` thinks the file was deleted.

fixes #487