Add aliases for built-in linters (#563)
These are equivalent to the aliases for the built-in formatters, and
allow use of these linters without having to use confusing relative repo
names that appear to be relative to the user's module, but are actually
relative to rules_lint.

Ideally the aliases would be `ruff` and `shellcheck`, but those names
are taken by bzl_libraries, hence the 'bin' suffix.

Fixes https://github.com/aspect-build/rules_lint/issues/551.