fix: remove dep on aspect_bazel_lib (#668)
We only want to constrain the version *if* some other module depends on
it. That's what the "nodep" feature does.

We should probably use this for some of our language dependencies as
well. That way rules_lint doesn't cause users to gain a dependency, for
example `aspect_rules_js` when they have zero JS/TS code.