feat: Implement clippy support  (#667)
Adds clippy support to rules_lint, which is part of #385. It does not
support `--fix` yet, as there are some complications with applying the
patches generated by rules_rust.

Caveats:
- `fix` doesn't work.
- The machine outputs are in in clippy's native JSON, not SARIF. We
could use cargo-sarif to transform it, but we haven't for now.
- We treat all clippy warnings as errors. This is because `rules_rust`
will fail the entire execution if there's even one error, so we need to
limit the reports to just warnings so that we can continue the target
execution and generate useful output files. Because we limit all errors
to warnings, we must consider every warning as an error.

---

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config):
Only if they were depending on rules_rust before, it may change their
`MODULE.lock` resolution.
- Suggested release notes appear below: yes/no

### Test plan

- New test cases added
- Some of the cases are manual: `cd examples && bazel test
//test:clippy_ok_binary_with_bad_dep //test:clippy_bad_binary
//test:clippy_bad_library`

---------

Co-authored-by: Alex Eagle <alex@aspect.dev>