feat: Add ty as a linter (#665)
Adds [ty](https://github.com/astral-sh/ty) as a linter for Python.

This is my first contribution here, so a lot of this is copied from the
implementation for `ruff`, including the strategy for mirroring `ty`
releases.

### Changes are visible to end-users: yes

- Searched for relevant documentation and updated as needed: yes
- Breaking change (forces users to change their own code or config): no
- Suggested release notes appear below: no

### Test plan

- New test cases added

Added `unsupported_operator.py` as an example lint, and
`call_non_callable.py` , which should be ignored by the `ty.toml`
configuration file.

---------

Co-authored-by: Alex Eagle <alex@aspect.dev>