refactor: split up giant monorepo example (#679)
It has reached a tipping point where some amount of duplication is worth
it to show something easier for users to digest.

As this targets the 2.x branch, also take the opportunities:
- remove `lint.sh` "workaround" since the AXL variant is better
supported. I'll move it to a public gist so it's still discoverable from
docs, but no longer maintained by Aspect
- replace linter registrations from old .aspect/cli/config.yaml to
.bazelrc files

In a follow-up PR I'll actually delete the /example folder.